All the bugs in delete frame seem to be fixed.
Next, apply delete frame code to model, color, and other tables.

+++ BikeShop/cSharp/bikes.data/ADO/FrameRepoADO.cs
@@ -98,7 +98,7 @@ namespace bikes.data.ADO
             return BikesWithFrame;
         }
 
-        public static void Delete(short FrameIdToDelete)
+        public static void Delete(int FrameIdToDelete)
         {
             using (var cn = new SqlConnection(Settings.GetConnectionString()))
             {
